+++ docs/conf.py
@@ -40,6 +40,8 @@ extensions = [
     'sphinx.ext.intersphinx',
     'sphinx.ext.autosectionlabel',
     'sphinx.ext.mathjax', # equations
+    'sphinx.ext.graphviz',
+    'sphinx.ext.inheritance_diagram',
     'recommonmark',
     'sphinx_rtd_theme'
 ]
@@ -106,3 +108,9 @@ html_theme_options = {
 #html_static_path = ['_static']
 
 html_logo = '../Logo-NenuFAR-noir.svg'
+# html_logo = '../nenupy_logo.svg'
+
+# The name of an image file (within the static path) to use as favicon of the
+# docs.  This file should be a Windows icon file (.ico) being 16x16 or 32x32
+# pixels large.
+#html_favicon = None