+++ nenupy/beam/__init__.py
@@ -2,6 +2,12 @@
 # -*- coding: utf-8 -*-
 
 
+"""
+    .. inheritance-diagram:: nenupy.beam.beam nenupy.beam.hpxbeam
+        :parts: 3
+"""
+
+
 from .beam import Beam, ABeam, DBeam
 from .hpxbeam import HpxABeam, HpxDBeam