+++ nenupy/beamlet/__init__.py
@@ -2,6 +2,12 @@
 # -*- coding: utf-8 -*-
 
 
+"""
+    .. inheritance-diagram:: nenupy.beamlet.sdata nenupy.beamlet.beamlet nenupy.beamlet.bstdata nenupy.beamlet.sstdata
+        :parts: 3
+"""
+
+
 from .sdata import SData
 from .beamlet import Beamlet
 from .bstdata import BST_Data