+++ nenupy/crosslet/__init__.py
@@ -2,6 +2,12 @@
 # -*- coding: utf-8 -*-
 
 
+"""
+    .. inheritance-diagram:: nenupy.crosslet.xstdata nenupy.crosslet.tvdata
+        :parts: 3
+"""
+
+
 from .uvw import UVW
 from .crosslet import Crosslet
 from .xstdata import XST_Data