+++ nenupy/simulation/__init__.py
@@ -2,4 +2,11 @@
 # -*- coding: utf-8 -*-
 
 
+"""
+    .. inheritance-diagram:: nenupy.simulation.hpxsimu
+        :parts: 3
+"""
+
+
 from .hpxsimu import HpxSimu
+