+++ nenupy/skymodel/__init__.py
@@ -2,6 +2,12 @@
 # -*- coding: utf-8 -*-
 
 
+"""
+    .. inheritance-diagram:: nenupy.skymodel.hpxgsm nenupy.skymodel.hpxlofar nenupy.skymodel.pointsource
+        :parts: 3
+"""
+
+
 from .hpxgsm import HpxGSM
 from .hpxlofar import HpxLOFAR
 from .pointsource import *