+++ nenupy/undysputed/__init__.py
@@ -1,6 +1,13 @@
 #! /usr/bin/python3
 # -*- coding: utf-8 -*-
 
+
+"""
+    .. inheritance-diagram:: nenupy.undysputed.dynspec nenupy.undysputed.waveform
+        :parts: 3
+"""
+
+
 from .dynspec import Dynspec
 from .waveform import Waveform